refactor: SessionUtils 의 메소드 AuthenticationUtils 클래스로 이동
SessionUtils 안에서 authentication 관련된 메소드는 AuthenticationUtils 로 이동
